initial commit 1

+++ app/src/main/java/com/example/weatherforcasteapp/POJO/Sys.kt
@@ -6,6 +6,6 @@ data class Sys(
     @SerializedName ("type") val type : Int,
     @SerializedName ("id") val id : Int,
     @SerializedName ("country") val country : String,
-    @SerializedName("sunrise") val sunrise :Int,
-    @SerializedName("sunset") val sunset :Int
+    @SerializedName("sunrise") val sunrise :Long,
+    @SerializedName("sunset") val sunset :Long
 )
